+++ src/python/magnum/platform/glfw.cpp
@@ -196,6 +196,24 @@ void glfw(py::module_& m) {
             }, "Swap interval")
         .def("main_loop_iteration", &PyApplication::mainLoopIteration, "Run one iteration of application main loop");
 
+    py::class_<Platform::Application::Configuration> configuration_{glfwApplication, "Configuration", "Configuration"};
+
+    py::enum_<Platform::Application::Configuration::WindowFlag> configurationWindowFlags{configuration_, "WindowFlags", "Window flags"};
+    configurationWindowFlags
+        .value("FULLSCREEN", Platform::Application::Configuration::WindowFlag::Fullscreen)
+        .value("BORDERLESS", Platform::Application::Configuration::WindowFlag::Borderless)
+        .value("RESIZABLE", Platform::Application::Configuration::WindowFlag::Resizable)
+        .value("HIDDEN", Platform::Application::Configuration::WindowFlag::Hidden)
+        .value("MAXIMIZED", Platform::Application::Configuration::WindowFlag::Maximized)
+        .value("MINIMIZED", Platform::Application::Configuration::WindowFlag::Minimized)
+        .value("ALWAYS_ON_TOP", Platform::Application::Configuration::WindowFlag::AlwaysOnTop)
+        .value("AUTO_ICONIFY", Platform::Application::Configuration::WindowFlag::AutoIconify)
+        .value("FOCUSED", Platform::Application::Configuration::WindowFlag::Focused)
+        /** @todo Contextless, once anything else than GL is exposed to
+            Python */
+        .value("NONE", Platform::Application::Configuration::WindowFlag{});
+    corrade::enumOperators(configurationWindowFlags);
+
     PyNonDestructibleClass<PublicizedApplication::ExitEvent> exitEvent_{glfwApplication, "ExitEvent", "Exit event"};
     PyNonDestructibleClass<PublicizedApplication::ViewportEvent> viewportEvent_{glfwApplication, "ViewportEvent", "Viewport event"};
     PyNonDestructibleClass<PublicizedApplication::InputEvent> inputEvent_{glfwApplication, "InputEvent", "Base for input events"};
@@ -234,6 +252,7 @@ void glfw(py::module_& m) {
         .value("HIDDEN", Platform::Application::Cursor::Hidden)
         .value("HIDDEN_LOCKED", Platform::Application::Cursor::HiddenLocked);
 
+    configuration(configuration_);
     application(glfwApplication);
     exitEvent(exitEvent_);
     viewportEvent(viewportEvent_);